requirements.txt
category-encoders==2.6.3
dataclasses==0.6
fqdn==1.5.1
fuzzywuzzy==0.18.0
geopy==2.4.1
importlib-metadata==8.4.0
importlib-resources==6.4.0
isoduration==20.11.0
jaraco.text==3.12.1
kneed==0.8.5
langchain==0.2.16
langchain-openai==0.1.23
notebook==7.2.2
ordered-set==4.1.0
pandasai==2.2.14
pickleshare==0.7.5
plotly==5.24.1
python-levenshtein==0.25.1
sentence-transformers==3.0.1
sqlglotrs==0.2.12
streamlit==1.38.0
tzdata==2024.1
uri-template==1.3.0
webcolors==24.8.0

-e .